feat(album): add hooks for album management and viewer functionality
- Implemented `useAlbumDrag` for drag-and-drop file import handling.
- Created `useAlbumGridConfig` for managing album grid configuration including columns, sorting, filtering, and grouping.
- Developed `useAlbumItems` for album item management with CRUD operations, metadata loading, and persistence.
- Added `useAlbumSelection` for multi-selection management in albums.
- Introduced `useAlbumViewer` for single image/video viewing with zoom, pan, and video controls.
- Implemented `useAlbumViewerActions` for toolbar actions like opening in explorer, saving images, and setting backgrounds.
